Replace chars and shorts with ints to improve performance

+++ renderer.h
@@ -2,18 +2,18 @@
 
 u16* g_buffer = (u16*)MODE5_BB;
 
-static u8 left_edges[SCREEN_HEIGHT];
-static u8 right_edges[SCREEN_HEIGHT];
-static u8 top_edge;
-static u8 bottom_edge;
+static u32 left_edges[SCREEN_HEIGHT];
+static u32 right_edges[SCREEN_HEIGHT];
+static u32 top_edge;
+static u32 bottom_edge;
 
-static void swap(u16* a, u16* b) {
-  u16 tmp = *a;
+static void swap(s32* a, s32* b) {
+  s32 tmp = *a;
   *a = *b;
   *b = tmp;
 }
 
-void clear_screen(u8 color) {
+void clear_screen(u32 color) {
   u32 value = (color << 24) | (color << 16) | (color << 8) | color;
 
   for (u32 i = 0; i < SCREEN_WIDTH * SCREEN_HEIGHT / 4; i++) {
@@ -21,9 +21,9 @@ void clear_screen(u8 color) {
   }
 }
 
-void put_pixel(u8 x, u8 y, u8 color) {
+void put_pixel(u32 x, u32 y, u32 color) {
   u32 index = (y * SCREEN_WIDTH / 2) + (x / 2);
-  u16 pair = g_buffer[index];
+  u32 pair = g_buffer[index];
 
   if (x & 1) {
     g_buffer[index] = (pair & 0xFF) | (color << 8);
@@ -32,14 +32,14 @@ void put_pixel(u8 x, u8 y, u8 color) {
   }
 }
 
-void trace_edge(u16 x0, u16 y0, u16 x1, u16 y1, u8* buffer) {
+void trace_edge(s32 x0, s32 y0, s32 x1, s32 y1, u32* buffer) {
   if (y0 > y1) {
     swap(&x0, &x1);
     swap(&y0, &y1);
   }
 
-  u8 start_y = y0 >> 8;
-  u8 end_y = y1 >> 8;
+  u32 start_y = y0 >> 8;
+  u32 end_y = y1 >> 8;
 
   if (start_y < top_edge) {
     top_edge = start_y;
@@ -58,7 +58,7 @@ void trace_edge(u16 x0, u16 y0, u16 x1, u16 y1, u8* buffer) {
   }
 }
 
-void draw_polygon(Point* points, u32 point_count, u8 color) {
+void draw_polygon(Point* points, u32 point_count, u32 color) {
   top_edge = 255;
   bottom_edge = 0;
 
@@ -74,7 +74,7 @@ void draw_polygon(Point* points, u32 point_count, u8 color) {
   }
 
   for (u32 y = top_edge; y < bottom_edge; y++) {
-    for (u8 x = left_edges[y]; x < right_edges[y]; x++) {
+    for (u32 x = left_edges[y]; x < right_edges[y]; x++) {
       put_pixel(x, y, 1);
     }
   }
